write_orbits_pd takes uint32_t instead of uint64_t array as image

+++ include/brot/iterate.h
@@ -18,7 +18,7 @@ int32_t write_orbits_ps(const float* cr, const float* ci, uint64_t size,
     uint32_t* img, uint32_t width, uint32_t height);
 int64_t write_orbits_pd(const double* cr, const double* ci, uint64_t size,
     uint64_t maxiter, double min_r, double max_r, double min_i, double max_i,
-    uint64_t* img, uint64_t width, uint64_t height);
+    uint32_t* img, uint64_t width, uint64_t height);
 
 #ifdef __cplusplus
 }
